Add OCMEM support
The OCMEM driver handles allocation and configuration of the On Chip
MEMory that is present on some snapdragon devices.

Devices which have OCMEM do not have GMEM inside the gpu core, so the
gpu must instead use OCMEM to be functional.  Since currently the gpu
is the only OCMEM user with an upstream driver, this is just a minimal
implementation sufficient for statically allocating to the gpu it's
chunk of OCMEM.